04- Transcribe-English-Videos/requirements.txt
gradio
openai-whisper
setuptools
moviepy
